[chore] 총알 데미지를 int에서 float으로 변경 + Player가 총을 들고 걸어다닐 때 총구 방향이 반대로 휘어지는 오류 수정

+++ Team3_NecroByte/Assets/Scripts/Enemy.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int maxHealth;
-    public int curHealth;
+    public float maxHealth;
+    public float curHealth;
 
     Rigidbody rigid;
     BoxCollider boxCollider;
